Refactor Request into smaller components (#519)
* refactor into smaller components

* stray component

* Update packages/extension-ui/src/Popup/Signing/Request/SignArea.tsx


* lint
