+++ packages/extension-ui/src/Popup/Signing/Request/index.tsx
@@ -1,0 +1,140 @@
+// Copyright 2019-2020 @polkadot/extension-ui authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { ExtrinsicPayload } from '@polkadot/types/interfaces';
+import { AccountJson, RequestSign } from '@polkadot/extension-base/background/types';
+import { TypeRegistry } from '@polkadot/types';
+import { SignerPayloadJSON, SignerPayloadRaw } from '@polkadot/types/types';
+
+import React, { useCallback, useEffect, useContext, useState } from 'react';
+
+import { ActionContext, Address, VerticalSpace } from '../../../components';
+import { approveSignSignature } from '../../../messaging';
+import Bytes from '../Bytes';
+import Extrinsic from '../Extrinsic';
+import Qr from '../Qr';
+import SignArea from './SignArea';
+
+interface Props {
+  account: AccountJson;
+  buttonText: string;
+  isFirst: boolean;
+  request: RequestSign;
+  signId: string;
+  url: string;
+}
+
+interface Data {
+  hexBytes: string | null;
+  payload: ExtrinsicPayload | null;
+}
+
+// keep it global, we can and will re-use this across requests
+const registry = new TypeRegistry();
+
+function isRawPayload (payload: SignerPayloadJSON | SignerPayloadRaw): payload is SignerPayloadRaw {
+  return !!(payload as SignerPayloadRaw).data;
+}
+
+export default function Request ({ account: { isExternal }, buttonText, isFirst, request, signId, url }: Props): React.ReactElement<Props> | null {
+  const onAction = useContext(ActionContext);
+  const [{ hexBytes, payload }, setData] = useState<Data>({ hexBytes: null, payload: null });
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect((): void => {
+    const payload = request.payload;
+
+    if (isRawPayload(payload)) {
+      setData({
+        hexBytes: payload.data,
+        payload: null
+      });
+    } else {
+      registry.setSignedExtensions(payload.signedExtensions);
+
+      setData({
+        hexBytes: null,
+        payload: registry.createType('ExtrinsicPayload', payload, { version: payload.version })
+      });
+    }
+  }, [request]);
+
+  const _onSignature = useCallback(
+    ({ signature }: { signature: string }): Promise<void> =>
+      approveSignSignature(signId, signature)
+        .then(() => onAction())
+        .catch((error: Error): void => {
+          setError(error.message);
+          console.error(error);
+        }),
+    [onAction, signId]
+  );
+
+  if (payload !== null) {
+    const json = request.payload as SignerPayloadJSON;
+
+    return (
+      <>
+        <div>
+          <Address
+            address={json.address}
+            genesisHash={json.genesisHash}
+            isExternal={isExternal}
+          />
+        </div>
+        {isExternal
+          ? (
+            <Qr
+              onSignature={_onSignature}
+              payload={payload}
+              request={json}
+            />
+          ) : (
+            <Extrinsic
+              isDecoded={true}
+              payload={payload}
+              request={json}
+              url={url}
+            />
+          )
+        }
+        <SignArea
+          buttonText={buttonText}
+          error={error}
+          isExternal={isExternal}
+          isFirst={isFirst}
+          setError={setError}
+          signId={signId}
+        />
+      </>
+    );
+  } else if (hexBytes !== null) {
+    const raw = request.payload as SignerPayloadRaw;
+
+    return (
+      <>
+        <div>
+          <Address
+            address={raw.address}
+            isExternal={isExternal}
+          />
+        </div>
+        <Bytes
+          bytes={raw.data}
+          url={url}
+        />
+        <VerticalSpace />
+        <SignArea
+          buttonText={buttonText}
+          error={error}
+          isExternal={isExternal}
+          isFirst={isFirst}
+          setError={setError}
+          signId={signId}
+        />
+      </>
+    );
+  }
+
+  return null;
+}
